[adjust - logic - complet] colocando a tela home page no diretorio de telas
